+++ server/src/modes/script/javascript.ts
@@ -1,7 +1,6 @@
 import { LanguageModelCache, getLanguageModelCache } from '../../embeddedSupport/languageModelCache';
 import {
   SymbolInformation,
-  SymbolKind,
   CompletionItem,
   Location,
   SignatureHelp,
@@ -23,7 +22,10 @@ import {
   Position,
   FormattingOptions,
   DiagnosticTag,
-  MarkupContent
+  MarkupContent,
+  CodeAction,
+  CodeActionKind,
+  WorkspaceEdit
 } from 'vscode-languageserver-types';
 import { LanguageMode } from '../../embeddedSupport/languageModes';
 import { VueDocumentRegions, LanguageRange } from '../../embeddedSupport/embeddedSupport';
@@ -41,6 +43,7 @@ import { getComponentInfo } from './componentInfo';
 import { DependencyService, T_TypeScript, State } from '../../services/dependencyService';
 import { RefactorAction } from '../../types';
 import { IServiceHost } from '../../services/typescriptService/serviceHost';
+import { toCompletionItemKind, toSymbolKind } from '../../services/typescriptService/util';
 
 // Todo: After upgrading to LS server 4.0, use CompletionContext for filtering trigger chars
 // https://microsoft.github.io/language-server-protocol/specification#completion-request-leftwards_arrow_with_hook
@@ -124,7 +127,7 @@ export async function getJavascriptMode(
         // so we can safely cast diag to TextSpan
         return <Diagnostic>{
           range: convertRange(scriptDoc, diag as ts.TextSpan),
-          severity: DiagnosticSeverity.Error,
+          severity: convertTSDiagnosticCategoryToDiagnosticSeverity(diag.category),
           message: tsModule.flattenDiagnosticMessageText(diag.messageText, '\n'),
           tags,
           code: diag.code,
@@ -160,11 +163,14 @@ export async function getJavascriptMode(
           return {
             uri: doc.uri,
             position,
+            preselect: entry.isRecommended ? true : undefined,
             label,
             detail,
+            filterText: getFilterText(entry.insertText),
             sortText: entry.sortText + index,
-            kind: convertKind(entry.kind),
+            kind: toCompletionItemKind(entry.kind),
             textEdit: range && TextEdit.replace(range, entry.name),
+            insertText: entry.insertText,
             data: {
               // data used for resolving item details (see 'doResolve')
               languageId: scriptDoc.languageId,
@@ -339,7 +345,7 @@ export async function getJavascriptMode(
         if (item.kind !== 'script' && !existing[sig]) {
           const symbol: SymbolInformation = {
             name: item.text,
-            kind: convertSymbolKind(item.kind),
+            kind: toSymbolKind(item.kind),
             location: {
               uri: doc.uri,
               range: convertRange(scriptDoc, item.spans[0])
@@ -435,7 +441,7 @@ export async function getJavascriptMode(
 
       const formatSettings: ts.FormatCodeSettings = getFormatCodeSettings(config);
 
-      const result: Command[] = [];
+      const result: CodeAction[] = [];
       const fixes = service.getCodeFixesAtPosition(
         fileName,
         start,
@@ -452,7 +458,7 @@ export async function getJavascriptMode(
 
       return result;
     },
-    getRefactorEdits(doc: TextDocument, args: RefactorAction) {
+    getRefactorEdits(doc: TextDocument, args: RefactorAction): WorkspaceEdit {
       const { service } = updateCurrentVueTextDocument(doc);
       const response = service.getEditsForRefactor(
         args.fileName,
@@ -464,10 +470,9 @@ export async function getJavascriptMode(
       );
       if (!response) {
         // TODO: What happens when there's no response?
-        return createApplyCodeActionCommand('', {});
+        return {};
       }
-      const uriMapping = createUriMappingForEdits(response.edits, service);
-      return createApplyCodeActionCommand('', uriMapping);
+      return { changes: createUriMappingForEdits(response.edits, service) };
     },
     format(doc: TextDocument, range: Range, formatParams: FormattingOptions): TextEdit[] {
       const { scriptDoc, service } = updateCurrentVueTextDocument(doc);
@@ -481,7 +486,7 @@ export async function getJavascriptMode(
         return [];
       }
 
-      const parser = scriptDoc.languageId === 'javascript' ? 'babylon' : 'typescript';
+      const parser = scriptDoc.languageId === 'javascript' ? 'babel' : 'typescript';
       const needInitialIndent = config.vetur.format.scriptInitialIndent;
       const vlsFormatConfig: VLSFormatConfig = config.vetur.format;
 
@@ -551,7 +556,7 @@ function collectRefactoringCommands(
   fileName: string,
   formatSettings: any,
   textRange: { pos: number; end: number },
-  result: Command[]
+  result: CodeAction[]
 ) {
   const actions: RefactorAction[] = [];
   for (const refactoring of refactorings) {
@@ -582,9 +587,13 @@ function collectRefactoringCommands(
   }
   for (const action of actions) {
     result.push({
-      command: 'vetur.chooseTypeScriptRefactoring',
       title: action.description,
-      arguments: [action]
+      kind: CodeActionKind.Refactor,
+      command: {
+        title: action.description,
+        command: 'vetur.chooseTypeScriptRefactoring',
+        arguments: [action]
+      }
     });
   }
 }
@@ -592,23 +601,25 @@ function collectRefactoringCommands(
 function collectQuickFixCommands(
   fixes: ReadonlyArray<ts.CodeFixAction>,
   service: ts.LanguageService,
-  result: Command[]
+  result: CodeAction[]
 ) {
   for (const fix of fixes) {
     const uriTextEditMapping = createUriMappingForEdits(fix.changes, service);
-    result.push(createApplyCodeActionCommand(fix.description, uriTextEditMapping));
+    result.push(createApplyCodeAction(CodeActionKind.QuickFix, fix.description, uriTextEditMapping));
   }
 }
 
-function createApplyCodeActionCommand(title: string, uriTextEditMapping: Record<string, TextEdit[]>): Command {
+function createApplyCodeAction(
+  kind: CodeActionKind,
+  title: string,
+  uriTextEditMapping: Record<string, TextEdit[]>
+): CodeAction {
   return {
     title,
-    command: 'vetur.applyWorkspaceEdits',
-    arguments: [
-      {
-        changes: uriTextEditMapping
-      }
-    ]
+    kind,
+    edit: {
+      changes: uriTextEditMapping
+    }
   };
 }
 
@@ -646,70 +657,6 @@ function convertRange(document: TextDocument, span: ts.TextSpan): Range {
   const startPosition = document.positionAt(span.start);
   const endPosition = document.positionAt(span.start + span.length);
   return Range.create(startPosition, endPosition);
-}
-
-function convertKind(kind: ts.ScriptElementKind): CompletionItemKind {
-  switch (kind) {
-    case 'primitive type':
-    case 'keyword':
-      return CompletionItemKind.Keyword;
-    case 'var':
-    case 'local var':
-      return CompletionItemKind.Variable;
-    case 'property':
-    case 'getter':
-    case 'setter':
-      return CompletionItemKind.Field;
-    case 'function':
-    case 'method':
-    case 'construct':
-    case 'call':
-    case 'index':
-      return CompletionItemKind.Function;
-    case 'enum':
-      return CompletionItemKind.Enum;
-    case 'module':
-      return CompletionItemKind.Module;
-    case 'class':
-      return CompletionItemKind.Class;
-    case 'interface':
-      return CompletionItemKind.Interface;
-    case 'warning':
-      return CompletionItemKind.File;
-    case 'script':
-      return CompletionItemKind.File;
-    case 'directory':
-      return CompletionItemKind.Folder;
-  }
-
-  return CompletionItemKind.Property;
-}
-
-function convertSymbolKind(kind: ts.ScriptElementKind): SymbolKind {
-  switch (kind) {
-    case 'var':
-    case 'local var':
-    case 'const':
-      return SymbolKind.Variable;
-    case 'function':
-    case 'local function':
-      return SymbolKind.Function;
-    case 'enum':
-      return SymbolKind.Enum;
-    case 'module':
-      return SymbolKind.Module;
-    case 'class':
-      return SymbolKind.Class;
-    case 'interface':
-      return SymbolKind.Interface;
-    case 'method':
-      return SymbolKind.Method;
-    case 'property':
-    case 'getter':
-    case 'setter':
-      return SymbolKind.Property;
-  }
-  return SymbolKind.Variable;
 }
 
 function convertOptions(
@@ -765,4 +712,42 @@ function convertCodeAction(
     }
   }
   return textEdits;
+}
+
+function convertTSDiagnosticCategoryToDiagnosticSeverity(c: ts.DiagnosticCategory) {
+  switch (c) {
+    case ts.DiagnosticCategory.Error:
+      return DiagnosticSeverity.Error;
+    case ts.DiagnosticCategory.Warning:
+      return DiagnosticSeverity.Warning;
+    case ts.DiagnosticCategory.Message:
+      return DiagnosticSeverity.Information;
+    case ts.DiagnosticCategory.Suggestion:
+      return DiagnosticSeverity.Hint;
+  }
+}
+
+/* tslint:disable:max-line-length */
+/**
+ * Adapted from https://github.com/microsoft/vscode/blob/2b090abd0fdab7b21a3eb74be13993ad61897f84/extensions/typescript-language-features/src/languageFeatures/completions.ts#L147-L181
+ */
+function getFilterText(insertText: string | undefined): string | undefined {
+  // For `this.` completions, generally don't set the filter text since we don't want them to be overly prioritized. #74164
+  if (insertText?.startsWith('this.')) {
+    return undefined;
+  }
+
+  // Handle the case:
+  // ```
+  // const xyz = { 'ab c': 1 };
+  // xyz.ab|
+  // ```
+  // In which case we want to insert a bracket accessor but should use `.abc` as the filter text instead of
+  // the bracketed insert text.
+  else if (insertText?.startsWith('[')) {
+    return insertText.replace(/^\[['"](.+)[['"]\]$/, '.$1');
+  }
+
+  // In all other cases, fallback to using the insertText
+  return insertText;
 }