+++ server/src/modes/template/services/htmlCompletion.ts
@@ -13,10 +13,7 @@ import { HTMLDocument } from '../parser/htmlParser';
 import { TokenType, createScanner, ScannerState } from '../parser/htmlScanner';
 import { IHTMLTagProvider } from '../tagProviders';
 import * as emmet from 'vscode-emmet-helper';
-import { VueFileInfo } from '../../../services/vueInfoService';
-import { doVueInterpolationComplete } from './vueInterpolationCompletion';
 import { NULL_COMPLETION } from '../../nullMode';
-import { isInsideInterpolation } from './isInsideInterpolation';
 import { getModifierProvider, Modifier } from '../modifierProvider';
 import { toMarkupContent } from '../../../utils/strings';
 
@@ -25,8 +22,7 @@ export function doComplete(
   position: Position,
   htmlDocument: HTMLDocument,
   tagProviders: IHTMLTagProvider[],
-  emmetConfig: emmet.EmmetConfiguration,
-  vueFileInfo?: VueFileInfo
+  emmetConfig: emmet.EmmetConfiguration
 ): CompletionList {
   const modifierProvider = getModifierProvider();
 
@@ -37,28 +33,8 @@ export function doComplete(
 
   const offset = document.offsetAt(position);
   const node = htmlDocument.findNodeBefore(offset);
-  if (!node) {
+  if (!node || node.isInterpolation) {
     return result;
-  }
-
-  const nodeRange = Range.create(document.positionAt(node.start), document.positionAt(node.end));
-  const nodeText = document.getText(nodeRange);
-  const insideInterpolation = isInsideInterpolation(node, nodeText, document.offsetAt(position) - node.start);
-
-  if (node.isInterpolation) {
-    if (!insideInterpolation) {
-      return NULL_COMPLETION;
-    }
-
-    if (document.getText()[document.offsetAt(position) - 1] === '.') {
-      return NULL_COMPLETION;
-    }
-
-    if (vueFileInfo) {
-      return doVueInterpolationComplete(vueFileInfo);
-    } else {
-      return NULL_COMPLETION;
-    }
   }
 
   const text = document.getText();
@@ -169,10 +145,9 @@ export function doComplete(
     const value = isFollowedBy(text, nameEnd, ScannerState.AfterAttributeName, TokenType.DelimiterAssign)
       ? ''
       : '="$1"';
-    const tag = currentTag.toLowerCase();
     tagProviders.forEach(provider => {
       const priority = provider.priority;
-      provider.collectAttributes(tag, (attribute, type, documentation) => {
+      provider.collectAttributes(currentTag, (attribute, type, documentation) => {
         if ((type === 'event' && filterPrefix !== '@') || (type !== 'event' && filterPrefix === '@')) {
           return;
         }
@@ -237,11 +212,7 @@ export function doComplete(
 
   function collectAttributeValueSuggestions(attr: string, valueStart: number, valueEnd?: number): CompletionList {
     if (attr.startsWith('v-') || attr.startsWith('@') || attr.startsWith(':')) {
-      if (vueFileInfo && insideInterpolation) {
-        return doVueInterpolationComplete(vueFileInfo);
-      } else {
-        return NULL_COMPLETION;
-      }
+      return NULL_COMPLETION;
     }
 
     let range: Range;
@@ -259,10 +230,9 @@ export function doComplete(
       range = getReplaceRange(valueStart, valueEnd);
       addQuotes = true;
     }
-    const tag = currentTag.toLowerCase();
     const attribute = currentAttributeName.toLowerCase();
     tagProviders.forEach(provider => {
-      provider.collectValues(tag, attribute, value => {
+      provider.collectValues(currentTag, attribute, value => {
         const insertText = addQuotes ? '"' + value + '"' : value;
         result.items.push({
           label: value,