+++ server/src/services/typescriptService/bridge.ts
@@ -28,7 +28,7 @@ export declare const ${componentHelperName}: {
   ): any;
 };
 export declare const ${iterationHelperName}: {
-  <T>(list: T[], fn: (value: T, index: number) => any): any;
+  <T>(list: readonly T[], fn: (value: T, index: number) => any): any;
   <T>(obj: { [key: string]: T }, fn: (value: T, key: string, index: number) => any): any;
   (num: number, fn: (value: number) => any): any;
   <T>(obj: object, fn: (value: any, key: string, index: number) => any): any;