+++ server/src/services/typescriptService/serviceHost.ts
@@ -14,6 +14,8 @@ import { TemplateSourceMap, stringifySourceMapNodes } from './sourceMap';
 import { isVirtualVueTemplateFile, isVueFile } from './util';
 import { logger } from '../../log';
 import { ModuleResolutionCache } from './moduleResolutionCache';
+import { globalScope } from './transformTemplate';
+import { inferVueVersion, VueVersion } from './vueVersion';
 
 const NEWLINE = process.platform === 'win32' ? '\r\n' : '\n';
 
@@ -406,7 +408,7 @@ export function getServiceHost(
 
   const registry = tsModule.createDocumentRegistry(true);
   let jsLanguageService = tsModule.createLanguageService(jsHost, registry);
-  const templateLanguageService = tsModule.createLanguageService(templateHost, registry);
+  const templateLanguageService = patchTemplateService(tsModule.createLanguageService(templateHost, registry));
 
   return {
     queryVirtualFileInfo,
@@ -415,6 +417,33 @@ export function getServiceHost(
     updateExternalDocument,
     dispose: () => {
       jsLanguageService.dispose();
+    }
+  };
+}
+
+function patchTemplateService(original: ts.LanguageService): ts.LanguageService {
+  const allowedGlobals = new Set(globalScope);
+
+  return {
+    ...original,
+
+    getCompletionsAtPosition(fileName, position, options) {
+      const result = original.getCompletionsAtPosition(fileName, position, options);
+      if (!result) {
+        return;
+      }
+
+      if (result.isMemberCompletion) {
+        return result;
+      }
+
+      return {
+        ...result,
+
+        entries: result.entries.filter(entry => {
+          return allowedGlobals.has(entry.name);
+        })
+      };
     }
   };
 }
@@ -436,50 +465,6 @@ function getScriptKind(tsModule: T_TypeScript, langId: string): ts.ScriptKind {
     : langId === 'tsx'
     ? tsModule.ScriptKind.TSX
     : tsModule.ScriptKind.JS;
-}
-
-enum VueVersion {
-  VPre25,
-  V25,
-  V30
-}
-
-function floatVersionToEnum(v: number) {
-  if (v < 2.5) {
-    return VueVersion.VPre25;
-  } else if (v < 3.0) {
-    return VueVersion.V25;
-  } else {
-    return VueVersion.V30;
-  }
-}
-
-function inferVueVersion(tsModule: T_TypeScript, workspacePath: string): VueVersion {
-  const packageJSONPath = tsModule.findConfigFile(workspacePath, tsModule.sys.fileExists, 'package.json');
-  try {
-    const packageJSON = packageJSONPath && JSON.parse(tsModule.sys.readFile(packageJSONPath)!);
-    const vueDependencyVersion = packageJSON.dependencies.vue || packageJSON.devDependencies.vue;
-
-    if (vueDependencyVersion) {
-      // use a sloppy method to infer version, to reduce dep on semver or so
-      const vueDep = vueDependencyVersion.match(/\d+\.\d+/)[0];
-      const sloppyVersion = parseFloat(vueDep);
-      return floatVersionToEnum(sloppyVersion);
-    }
-
-    const nodeModulesVuePackagePath = tsModule.findConfigFile(
-      path.resolve(workspacePath, 'node_modules/vue'),
-      tsModule.sys.fileExists,
-      'package.json'
-    );
-    const nodeModulesVuePackageJSON =
-      nodeModulesVuePackagePath && JSON.parse(tsModule.sys.readFile(nodeModulesVuePackagePath)!);
-    const nodeModulesVueVersion = parseFloat(nodeModulesVuePackageJSON.version.match(/\d+\.\d+/)[0]);
-
-    return floatVersionToEnum(nodeModulesVueVersion);
-  } catch (e) {
-    return VueVersion.VPre25;
-  }
 }
 
 function getParsedConfig(tsModule: T_TypeScript, workspacePath: string) {