+++ test/interpolation/fixture/completion/Parent.vue
@@ -2,16 +2,20 @@
   <div>
     <basic :></basic>
     <basic v-if="" @click="" :foo=""></basic>
+    <test-comp  />
+    <TestComp  />
     <
   </div>
 </template>
 
 <script>
 import Basic from './Basic.vue'
+import TestComp from './TestComp.vue'
 
 export default {
   components: {
-    Basic
+    Basic,
+    TestComp
   },
   props: {
     /**
@@ -35,7 +39,7 @@ export default {
      * My count
      */
     count () {
-      return this.$store.state.count
+      return 42;
     }
   },
   methods: {