+++ test/lsp/features/diagnostics/unused.test.ts
@@ -1,0 +1,27 @@
+import * as vscode from 'vscode';
+import { DiagnosticTag } from 'vscode-languageclient';
+import { getDocUri, sameLineRange } from '../../util';
+import { testDiagnostics } from './helper';
+
+describe('Should find diagnostics for unused variables', () => {
+  const docUri = getDocUri('diagnostics/Unused.vue');
+
+  it('shows diagnostic errors for unused variables', async () => {
+    const expectedDiagnostics: vscode.Diagnostic[] = [
+      {
+        severity: vscode.DiagnosticSeverity.Error,
+        message: "'lodash' is declared but its value is never read.",
+        range: sameLineRange(5, 0, 33),
+        tags: [DiagnosticTag.Unnecessary]
+      },
+      {
+        severity: vscode.DiagnosticSeverity.Error,
+        message: "'foo' is declared but its value is never read.",
+        range: sameLineRange(7, 6, 9),
+        tags: [DiagnosticTag.Unnecessary]
+      }
+    ];
+
+    await testDiagnostics(docUri, expectedDiagnostics);
+  });
+});