+++ test/lsp/fixture/completion/template/childComponent/Child1775.vue
@@ -1,0 +1,10 @@
+<template>
+  <div></div>
+</template>
+
+<script>
+// https://github.com/vuejs/vetur/issues/1775
+export default {
+  props: ['attr']
+};
+</script>