+++ test/lsp/fixture/completion/template/childComponent/Parent1775.vue
@@ -1,0 +1,13 @@
+<template>
+  <child1775 ></child1775>
+</template>
+
+<script>
+import Child1775 from './Child1775.vue'
+
+export default {
+  components: {
+    Child1775
+  }
+}
+</script>