+++ test/vue3/features/completion/basic.test.ts
@@ -1,0 +1,18 @@
+import { CompletionItemKind } from 'vscode';
+import { getDocUri, position } from '../../util';
+import { testCompletion } from './helper';
+
+describe('Vue 3 integration test', () => {
+  const fileUri = getDocUri('completion/Basic.vue');
+
+  describe('Should complete Vue 3 options', () => {
+    it('complete `setup`', async () => {
+      await testCompletion(fileUri, position(6, 2), [
+        {
+          label: 'setup',
+          kind: CompletionItemKind.Field
+        }
+      ]);
+    });
+  });
+});