+++ test/vue3/index.ts
@@ -1,9 +1,37 @@
-import * as testRunner from 'vscode/lib/testrunner';
+import * as path from 'path';
+import * as Mocha from 'mocha';
+import * as glob from 'glob';
 
-testRunner.configure({
-  ui: 'bdd',
-  useColors: true,
-  timeout: 100000
-});
+export function run(): Promise<void> {
+  // Create the mocha test
+  const mocha = new Mocha({
+    ui: 'bdd',
+    timeout: 100000,
+    color: true
+  });
+  const testsRoot = __dirname;
 
-module.exports = testRunner;
+  return new Promise((c, e) => {
+    glob('**/**.test.js', { cwd: testsRoot }, (err, files) => {
+      if (err) {
+        return e(err);
+      }
+
+      // Add files to the test suite
+      files.forEach(f => mocha.addFile(path.resolve(testsRoot, f)));
+
+      try {
+        // Run the mocha test
+        mocha.run(failures => {
+          if (failures > 0) {
+            e(new Error(`${failures} tests failed.`));
+          } else {
+            c();
+          }
+        });
+      } catch (err) {
+        e(err);
+      }
+    });
+  });
+}